+++ custom_components/meteo_lt/weather.py
@@ -45,6 +45,11 @@ async def async_setup_entry(
 class MeteoLtWeather(SingleCoordinatorWeatherEntity[MeteoLtCoordinator]):
     """Meteo.lt WeatherEntity implementation"""
 
+    _attr_native_precipitation_unit = UnitOfPrecipitationDepth.MILLIMETERS
+    _attr_native_pressure_unit = UnitOfPressure.HPA
+    _attr_native_temperature_unit = UnitOfTemperature.CELSIUS
+    _attr_native_wind_speed_unit = UnitOfSpeed.METERS_PER_SECOND
+
     def __init__(self, coordinator, nearest_place, config_entry):
         """__init__"""
         super().__init__(coordinator)